+++ contracts/liquid-collateral/Cargo.toml
@@ -18,15 +18,11 @@ library = []
 
 [dependencies]
 serde = { version = "1.0.202", default-features = false, features = ["derive"] }
-cosmwasm-std = { version = "1.5.0", features = ["staking"] }
+cosmwasm-std = { version = "1.5.0"}
 cosmwasm-schema = { workspace = true }
 cw-storage-plus = "1.1.0"
 thiserror = "1.0.50"
 osmosis-std = "0.25.0"
-cw20 = "1.1.2"
-cw20-base = "1.1.2"
-cw-orch = { workspace = true }
-hydro = { path = "../hydro", features = ["library"] }
 serde_json = "1.0.0"
 schemars = "0.8"
 [dev-dependencies]
